Added month start handling
There's a few days delay (typically 3) between the start of the new month, and Suez making the daily numbers available to the dashboard. This fork helps address that by checking for ERR responses from Suez and stepping the time back one day at a time until a complete data set is retrieved successfully.